requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.7
sqlparse==0.4.4
typing-extensions==4.8.0

djangorestframework~=3.14.0
requests~=2.31.0
python-dotenv~=1.0.0